ingest/requirements.txt
black==23.3.0
llama-index==0.5.27
python-dotenv==1.0.0